backend/requirements.txt
psycopg2-binary==2.9.9
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0 